+++ tests/unit-pure/esnext.async-iterator.find.js
@@ -19,11 +19,16 @@ QUnit.test('AsyncIterator#find', assert => {
   assert.throws(() => find.call(createIterator([1]), null), TypeError);
   assert.throws(() => find.call(createIterator([1]), {}), TypeError);
 
+  const counters = [];
+
   return find.call(createIterator([2, 3, 4]), it => it % 2).then(result => {
     assert.same(result, 3, 'basic functionality, +');
     return find.call(createIterator([1, 2, 3]), it => it === 4);
   }).then(result => {
     assert.same(result, undefined, 'basic functionality, -');
+    return find.call(createIterator([1, 2, 3]), (value, counter) => { counters.push(counter); return false; });
+  }).then(() => {
+    assert.arrayEqual(counters, [0, 1, 2], 'counter incremented');
     return find.call(createIterator([1]), function (arg, counter) {
       assert.same(this, STRICT_THIS, 'this');
       assert.same(arguments.length, 2, 'arguments length');